+++ frameworks/bridge/js_frontend/engine/quickjs/qjs_engine.cpp
@@ -470,7 +470,7 @@ void SetDomStyle(JSContext* ctx, JSValueConst fromMap, JsCommandDomElementOperat
         JS_FreeValue(ctx, val);
     }
     bool isIine = false;
-    for (int32_t i = 0; i < styles.size(); i++) {
+    for (int32_t i = 0; i < static_cast<int32_t>(styles.size()); i++) {
         std::string key = styles[i].first;
         std::string value = styles[i].second;
         if (key == "display" && value == "inline") {
